@@ -27,6 +27,7 @@
     "@tiptap/pm": "^3.0.7",
     "@tiptap/react": "^3.0.7",
     "@tiptap/starter-kit": "^3.0.7",
+    "better-auth": "^1.3.34",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "dot-env": "^0.0.1",
@@ -38,8 +39,8 @@
     "lucide-react": "^0.525.0",
     "mic-recorder-to-mp3": "^2.2.2",
     "ogl": "^1.0.11",
-    "react": "^19.1.0",
-    "react-dom": "^19.1.0",
+    "react": "^19.2.1",
+    "react-dom": "^19.2.1",
     "react-hotkeys-hook": "^5.1.0",
     "react-router": "^7.8.2",
     "socket.io-client": "^4.8.1",
