@@ -1,6 +1,6 @@
-const List = require('../models/list')
-const Post = require('../models/post')
-const User = require('../models/user')
+import List from '../models/list.js';
+import Post from '../models/post.js';
+import User from '../models/user.js';
 
 const getAllList = async (req, res) => {
     
@@ -82,21 +82,72 @@ const deleteFromList = async (req, res) => {
     
 }
 
-const deleteList = async (req, res) => {
+const updateList = async (req, res) => {
+    try {
+        const listId = req.params.listId
+        const userId = req.user._id
+        const { name } = req.body
 
-    const listId = req.params.listId
-    const userId = req.user._id
+        // Find the list first to check ownership
+        const list = await List.findById(listId);
+        if (!list) {
+            return res.status(404).json({ error: "List not found" });
+        }
 
-    // delete the list itself
-    const data = await List.findByIdAndDelete(listId)
+        // Check if user is the owner
+        if (list.user.toString() !== userId.toString()) {
+            return res.status(403).json({ error: "You can only edit your own lists" });
+        }
 
-    // also remove reference from user.lists
-    await User.findByIdAndUpdate(
-      userId,
-      { $pull: { lists: listId } }
-    )
+        // Validate name
+        if (!name || !name.trim()) {
+            return res.status(400).json({ error: "List name is required" });
+        }
 
-    res.status(200).json(data)
+        // Update the list
+        const updatedList = await List.findByIdAndUpdate(
+            listId,
+            { name: name.trim() },
+            { new: true }
+        ).populate("blogs").exec();
+
+        res.status(200).json(updatedList);
+    } catch (e) {
+        console.log(e);
+        res.status(500).json({ error: "Failed to update list", details: e.message });
+    }
 }
 
-module.exports = {getAllList, fetchList, createList, addToList, deleteFromList, deleteList}
+const deleteList = async (req, res) => {
+    try {
+        const listId = req.params.listId
+        const userId = req.user._id
+
+        // Find the list first to check ownership
+        const list = await List.findById(listId);
+        if (!list) {
+            return res.status(404).json({ error: "List not found" });
+        }
+
+        // Check if user is the owner
+        if (list.user.toString() !== userId.toString()) {
+            return res.status(403).json({ error: "You can only delete your own lists" });
+        }
+
+        // delete the list itself
+        const data = await List.findByIdAndDelete(listId)
+
+        // also remove reference from user.lists
+        await User.findByIdAndUpdate(
+          userId,
+          { $pull: { lists: listId } }
+        )
+
+        res.status(200).json(data)
+    } catch (e) {
+        console.log(e);
+        res.status(500).json({ error: "Failed to delete list", details: e.message });
+    }
+}
+
+export { getAllList, fetchList, createList, addToList, deleteFromList, updateList, deleteList };
