@@ -1,4 +1,4 @@
-const mongoose = require('mongoose')
+import mongoose from 'mongoose';
 
 const Schema = mongoose.Schema
 
@@ -7,12 +7,11 @@ const commentModel = Schema(
         blog: {type: Schema.Types.ObjectId, ref: "post"},
         comment: {type: String},
         user: {type: Schema.Types.ObjectId, ref:"user"},
+        replyTo: {type: Schema.Types.ObjectId, ref: "comment", default: null},
     },
     {
         timestamps: true
     }
 )
 
-const comment = mongoose.model("comment", commentModel)
-
-module.exports = comment
+export default mongoose.model("comment", commentModel)
