@@ -49,17 +49,33 @@ const CreateListModal = ({ setListModal, listModal, profile, setLists}: CreateLi
   }
 
   return (
-    <div className='absolute z-1 flex h-[100vh] w-[100vw] justify-center items-center backdrop-blur-xs' onClick={() => setListModal(!listModal)}>
-        <div className='w-[300px] h-[400px] bg-[#393E46] border-solid border-[1px] border-[#979797] flex flex-col gap-2 items-center justify-center p-2' onClick={(e) => e.stopPropagation()}>
-            <label className='hidden' id="list" htmlFor='list'></label>
-            <input 
-                name="list" 
-                id="list" 
-                value={listName ?? ""}
-                onChange={(event) => {setListName(event.target.value)}} 
-                placeholder='enter list name'
-            />
-            <button className="primary-btn" onClick={createList}><span>CREATE LIST!</span></button>
+    <div className='fixed z-50 flex h-[100vh] w-[100vw] justify-center items-center bg-black/60 backdrop-blur-sm' onClick={() => setListModal(!listModal)}>
+        <div className='w-[400px] bg-[#0f1926] border border-[#1f2735] rounded-lg flex flex-col gap-4 p-6 shadow-xl' onClick={(e) => e.stopPropagation()}>
+            <div className="flex items-center justify-between">
+              <h3 className="text-white font-semibold text-lg">Create Collection</h3>
+              <button 
+                onClick={() => setListModal(!listModal)}
+                className="text-[#9aa4bd] hover:text-white transition-colors"
+              >
+                ×
+              </button>
+            </div>
+            
+            <div className="flex flex-col gap-3">
+              <label htmlFor="list" className="text-[#9aa4bd] text-sm">Collection Name</label>
+              <input 
+                  name="list" 
+                  id="list" 
+                  value={listName ?? ""}
+                  onChange={(event) => {setListName(event.target.value)}} 
+                  placeholder='Enter collection name'
+                  className="w-full px-4 py-2 bg-[#121b2a] border border-[#1f2735] rounded-lg text-white placeholder-[#9aa4bd] focus:outline-none focus:border-[#31415f]"
+              />
+            </div>
+            
+            <button className="w-full primary-btn py-3 rounded-lg" onClick={createList}>
+              <span>CREATE COLLECTION</span>
+            </button>
         </div>
     </div>
   )
